h265nal: fix build
Tested:

Before:
```
$ cd build
$ make
...
[ 23%] Building CXX object src/CMakeFiles/h265nal.dir/h265_nal_unit_payload_parser.cc.o
src/h265_sei_parser.cc:200:38: warning: format specifies type 'int' but the argument has type 'SeiType' [clang-diagnostic-format]
  200 |   fprintf(outfp, "payload_type: %i", payload_type);
      |                                 ~~   ^~~~~~~~~~~~
      |                                      static_cast<int>( )
src/h265_sei_parser.cc:200:38: error: format specifies type 'int' but the argument has type 'SeiType' [-Werror,-Wformat]
  200 |   fprintf(outfp, "payload_type: %i", payload_type);
      |                                 ~~   ^~~~~~~~~~~~
      |                                      static_cast<int>( )
1 error generated.
...
```

After:
```
$ cd build
$ make
...
[ 23%] Building CXX object src/CMakeFiles/h265nal.dir/h265_nal_unit_payload_parser.cc.o
...
```

+++ src/h265_sei_parser.cc
@@ -197,7 +197,7 @@ void H265SeiMessageParser::SeiMessageState::fdump(FILE* outfp,
   indent_level = indent_level_incr(indent_level);
 
   fdump_indent_level(outfp, indent_level);
-  fprintf(outfp, "payload_type: %i", payload_type);
+  fprintf(outfp, "payload_type: %i", static_cast<int>(payload_type));
 
   fdump_indent_level(outfp, indent_level);
   fprintf(outfp, "payload_size: %u", payload_size);
